src/requirements.txt
pycurl==7.43.0.3
tornado==5.1.1
elasticsearch[async]==7.10.1
fastapi==0.63.0
uvicorn[standard]==0.13.3
python-multipart==0.0.5
python-jose[cryptography]==3.2.0
passlib[bcrypt]==1.7.4
aiofiles==0.6.0
python-dotenv==0.15.0

